+++ shardingsphere-sql-parser/shardingsphere-sql-parser-binder/src/main/java/org/apache/shardingsphere/sql/parser/binder/metadata/column/ColumnMetaDataLoader.java
@@ -15,9 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.underlying.common.metadata.column.loader;
-
-import org.apache.shardingsphere.sql.parser.binder.metadata.column.ColumnMetaData;
+package org.apache.shardingsphere.sql.parser.binder.metadata.column;
 
 import java.sql.Connection;
 import java.sql.ResultSet;
